Fix wxString::Format() assert with wx-3.0.

+++ diff-pdf.cpp
@@ -656,7 +656,7 @@ private:
             (
                 wxT("Page %d of %d; %d of them %s different, this page %s"),
                 m_cur_page + 1 /* humans prefer 1-based counting*/,
-                m_pages.size(),
+                (int)m_pages.size(),
                 m_diff_count,
                 m_diff_count == 1 ? wxT("is") : wxT("are"),
                 m_pages[m_cur_page] ? wxT("differs") : wxT("is unchanged")
